requirements.txt
streamlit==0.86.0
pandas==1.3.2
joblib==0.14.1
Pillow==8.4.0
scikit_learn==1.0.1
